+++ component/helpers.cmake
@@ -30,3 +30,4 @@ include("${CMAKE_CURRENT_LIST_DIR}/materialize.cmake")
 # Backend wrappers (create_cmake_* / create_meson_*).
 include("${CMAKE_CURRENT_LIST_DIR}/cmake/helpers.cmake")
 include("${CMAKE_CURRENT_LIST_DIR}/meson/helpers.cmake")
+include("${CMAKE_CURRENT_LIST_DIR}/factory.cmake")